feat: Create generic repost when cross-posting nip22 comment

+++ app/src/main/java/com/dluvian/voyage/data/room/entity/main/CommentMetaEntity.kt
@@ -23,14 +23,14 @@ import com.dluvian.voyage.data.event.ValidatedComment
 data class CommentMetaEntity(
     val eventId: EventIdHex,
     val parentId: EventIdHex?, // We don't support a and i parent tags
-    val parentKind: Int,
+    val parentKind: Int, // We save this to easily determine if parent is renderable
 ) {
     companion object {
         fun from(comment: ValidatedComment): CommentMetaEntity {
             return CommentMetaEntity(
                 eventId = comment.id,
                 parentId = comment.parentId,
-                parentKind = comment.parentKind
+                parentKind = comment.parentKind.toInt()
             )
         }
     }
